Added KMC overloading header, removed uniqueFlag, and removed rod array api.
